Log the server correlation id for /userinfo
I feel like there needs to be a deeper overhaul of how we log our XHR
requests, but this takes care of the immediate need of being able to trace
the /userinfo calls on the server.  This is especially important in
tracking down the sign-in issues that are still hit on occasion.