Update string to include live.com in the message
When I was playing around with the third-party cookie settings in Firefox,
it looked like live.com was set as an exception by default, so I decided
to take it out of the original error message about cookies being disabled.
But, it seems that cookies are only allowed on the sign in endpoint.  For
whatever reason, if you don't have the live.com domain added as an
exception, you are unable to sign out of the clipper.

So, I've added back in the need to add live.com to the list of exceptions.