Fix the handleSignIn() error object
In order to know if we should show the error description when sign in
errors occur, we were counting on the updateReason property being set on
the returning object. This was true until my last change, where we always
return a valid UserInfo (except in this one case). After my change, the
updateReason property was getting blown away and because of that, the
logic around when to show the error message could never be true.