Generate the bookmarklet version
In order to make it easier to update the version numbers, we need to get
rid of anything that's hard-coded or defined in code. The only remaining
place was the bookmarklet, so I am now generating the variable at build
time. Note that there is probably a better way of doing this, but it
works. :-)