Reduce the Edge languages in the package
I'm reducing the package locales down to en-us only.  This doesn't mean
that we are not supporting the other locales, we do. What it does is make
the store submission process a little easier.  We have the description
localized in a whole lot of languages, but there really isn't a good
reason to keep this list in sync, and in fact caused problems the last
time I submitted. To keep things simple and easier to submit, I'm making
this en-us only in the package.

At runtime we still continue to show the 100+ languages that the clipper
is localized in.