Remove the injected div in non-Chrome browsers
It looks like the change we made earlier to remove the injected div in
every page didn't make it to any browser except Chrome. This adds the same
logic to the Edge/Firefox manifest.json files and removes it completely
from Safari.

Note: Also cleaned up the Edge manifest.json to be a little more
consistent with the other Web Extension-based browsers.