Remove local/extension URLs from the DOM
Because we are sending up the DOM to be rendered on the service (for the
full page screenshot), we sometimes run into problems because of a URL
that is impossible to resolve, URLs like ones injected by another
extension.  This change basically strips out those URLs before the DOM is
sent up.

Oh, and because I really wanted to make a private method, I converted
DomUtils.ts to a true class, which had a little bit of a cascading effect.
I have a bigger change that I want to do which converts all of our
exported functions into classes, but that's for another day.