Remove base64 encoded binaries in <style>
It turns out that there are a lot of exensions in the world, and sometimes those extensions modify the DOM of the page (ours included). Generally we don't have a problem with this, but if an extension injects a large base64-encoded binary into the mix, then we run into problems.  This is mostly because the request size very quickly can be bloated beyond the max.  This change targets one particular place where this happens the most, in the <style> tag where someone may inject a custom font into the page.  When we detect this, we remove the style (from our copy of the DOM) before sending it up.